Partial continue functionality

+++ main.py
@@ -37,8 +37,7 @@ if __name__ == '__main__':
     for path in os.listdir(dir_path):
         if os.path.isfile(os.path.join(dir_path, path)):
             starting_chapter += 1
-    iteration = starting_chapter + 1
-    print('Starting chapter is: ', starting_chapter)
+    iteration = 0
 
     # Iterates through all the chapters until there are no more chapters
     while current_page is not None:
@@ -46,35 +45,38 @@ if __name__ == '__main__':
         # Get html of current page
         soup = bs4.BeautifulSoup(html.text, 'html.parser')
 
-        # Getting the images here
-        reader_container = soup.find("div", {"class": "container-chapter-reader"})
-        image_links = [x['src'] for x in reader_container.findChildren("img")]
+        if iteration >= starting_chapter:
+            # Getting the images here
+            reader_container = soup.find("div", {"class": "container-chapter-reader"})
+            image_links = [x['src'] for x in reader_container.findChildren("img")]
 
-        # Here we iterate through all the images of the current chapter
-        for i in tqdm(range(len(image_links)), desc=f"Downloading chapter {iteration}!"):
-            image_url = image_links[i]
-            file_name = f"{folder_name}/image_{i + 1}.jpg"
+            # Here we iterate through all the images of the current chapter
+            for i in tqdm(range(len(image_links)), desc=f"Downloading chapter {iteration + 1}!"):
+                image_url = image_links[i]
+                file_name = f"{folder_name}/image_{i + 1}.jpg"
 
-            # We update the headers to create a reference to the current chapter page
-            headers = random.choice(headers_list)
-            headers["Referer"] = current_page
+                # We update the headers to create a reference to the current chapter page
+                headers = random.choice(headers_list)
+                headers["Referer"] = current_page
 
-            with session.get(image_url, headers=headers) as response:
-                with open(file_name, "wb") as f:
-                    f.write(response.content)
+                with session.get(image_url, headers=headers) as response:
+                    with open(file_name, "wb") as f:
+                        f.write(response.content)
 
-            # We add it to the current image list and initialize the first image if it is not already
-            if pdf is None:
-                pdf = Image.open(file_name).convert('RGB')
-            else:
-                pdf_pages.append(Image.open(file_name).convert('RGB'))
+                # We add it to the current image list and initialize the first image if it is not already
+                if pdf is None:
+                    pdf = Image.open(file_name).convert('RGB')
+                else:
+                    pdf_pages.append(Image.open(file_name).convert('RGB'))
 
-            # We remove the current image from the pc as it is already stocked in the memory and sleep if we need to
-            os.remove(file_name)
-            time.sleep(delay_between_requests)
+                # We remove the current image from the pc as it is already stocked in the memory and sleep if we need to
+                os.remove(file_name)
+                time.sleep(delay_between_requests)
+            pdf.save(f'{folder_name}/chapter_{iteration + 1}.pdf', save_all=True, append_images=pdf_pages)
+        else:
+            print(f"Skipping chapter {iteration + 1}, it already exists!")
 
         # We check if there is a next chapter
-        pdf.save(f'{folder_name}/chapter_{iteration}.pdf', save_all=True, append_images=pdf_pages)
         next_chapter = soup.find("a", {"class": "navi-change-chapter-btn-next"})
         if next_chapter is not None:
             # If there is we go to the next page and create the pdf of the current one
@@ -86,6 +88,6 @@ if __name__ == '__main__':
             # If not, then we end the program
             current_page = None
 
-    print(f"Finished creating {iteration - starting_chapter} chapter pdfs!")
+    print(f"Finished creating {iteration - starting_chapter + 1} chapter pdfs!")
